+++ cuda/include/cuda/algorithms/seeding/detail/triplet_counter.hpp
@@ -14,7 +14,7 @@ namespace cuda {
 
 struct triplet_counter {
     doublet mid_bot_doublet;
-    size_t n_triplets = 0;
+    unsigned int n_triplets = 0;
 };
 
 /// Container of triplet_counter belonging to one detector module
